(maint) Objectify the validator classes
  * previously the validator classes were inheriting from a class that
    served several purposes. This moves that code into a similar class
    called validator_base that acts as the true parent so we can
    separate out concerns easier. This cleans up the code as it becomes
    easier to see what the parent is doing vs the validator module.

    Additional the validator class is now a module instead with a few
    methods that do things.
